// Client/src/app/core/modules/openapi/model/create-degree-program-specialization-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateDegreeProgramSpecializationDTO { 
    name: string;
    responsibleUserId: string;
}