// Client/src/app/core/modules/openapi/model/degree-program-specialization-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ResponsibleUserDTO } from './responsible-user-dto';


export interface DegreeProgramSpecializationDTO { 
    degreeProgramSpecializationId: number;
    name: string;
    responsibleUser: ResponsibleUserDTO;
}